+++ apps/client/src/app/components/user-account-access/create-or-update-access-dialog/create-or-update-access-dialog.component.ts
@@ -32,6 +32,7 @@ import {
   Validators
 } from '@angular/forms';
 import { MatButtonModule } from '@angular/material/button';
+import { MatDatepickerModule } from '@angular/material/datepicker';
 import {
   MAT_DIALOG_DATA,
   MatDialogModule,
@@ -40,7 +41,7 @@ import {
 import { MatFormFieldModule } from '@angular/material/form-field';
 import { MatInputModule } from '@angular/material/input';
 import { MatSelectModule } from '@angular/material/select';
-import { AccessPermission } from '@prisma/client';
+import { AccessPermission, AccessType } from '@prisma/client';
 import { StatusCodes } from 'http-status-codes';
 import { EMPTY, catchError } from 'rxjs';
 
@@ -53,6 +54,7 @@ import { CreateOrUpdateAccessDialogParams } from './interfaces/interfaces';
     FormsModule,
     GfPortfolioFilterFormComponent,
     MatButtonModule,
+    MatDatepickerModule,
     MatDialogModule,
     MatFormFieldModule,
     MatInputModule,
@@ -70,9 +72,11 @@ export class GfCreateOrUpdateAccessDialogComponent implements OnInit {
   public tags: Filter[] = [];
 
   protected accessForm: FormGroup;
+  protected hasExperimentalFeatures = false;
+  protected readonly minExpiresAtDate = new Date();
   protected readonly mode: 'create' | 'update';
 
-  private hasExperimentalFeatures = false;
+  private hasLoadedFilters = false;
 
   private readonly changeDetectorRef = inject(ChangeDetectorRef);
 
@@ -95,21 +99,27 @@ export class GfCreateOrUpdateAccessDialogComponent implements OnInit {
 
   public get canApplyFilters() {
     return (
-      this.accessForm?.get('type')?.value === 'PUBLIC' &&
+      ['API', 'PUBLIC'].includes(this.accessForm?.get('type')?.value) &&
       this.hasExperimentalFeatures
     );
   }
 
   public ngOnInit() {
     const access = this.data?.access;
-    const isPublic = access?.type === 'PUBLIC';
+    const isPrivate = (access?.type ?? 'PRIVATE') === 'PRIVATE';
 
     this.accessForm = this.formBuilder.group({
       alias: [access?.alias ?? ''],
+      expiresAt: [
+        {
+          disabled: this.mode === 'update',
+          value: access?.expiresAt ?? null
+        }
+      ],
       filters: [null],
       granteeUserId: [
         access?.grantee ?? null,
-        isPublic ? null : Validators.required
+        isPrivate ? Validators.required : null
       ],
       permissions: [
         access?.permissions[0] ?? AccessPermission.READ_RESTRICTED,
@@ -138,6 +148,7 @@ export class GfCreateOrUpdateAccessDialogComponent implements OnInit {
       .get('type')
       ?.valueChanges.pipe(takeUntilDestroyed(this.destroyRef))
       .subscribe((accessType) => {
+        const expiresAtControl = this.accessForm.get('expiresAt');
         const granteeUserIdControl = this.accessForm.get('granteeUserId');
         const permissionsControl = this.accessForm.get('permissions');
 
@@ -147,9 +158,18 @@ export class GfCreateOrUpdateAccessDialogComponent implements OnInit {
         } else {
           granteeUserIdControl?.clearValidators();
           granteeUserIdControl?.setValue(null);
-          permissionsControl?.setValue(
-            access?.permissions[0] ?? AccessPermission.READ_RESTRICTED
-          );
+
+          if (accessType === 'API') {
+            permissionsControl?.setValue(AccessPermission.READ);
+          } else {
+            permissionsControl?.setValue(
+              access?.permissions[0] ?? AccessPermission.READ_RESTRICTED
+            );
+          }
+        }
+
+        if (accessType !== 'API') {
+          expiresAtControl?.setValue(null);
         }
 
         granteeUserIdControl?.updateValueAndValidity();
@@ -181,8 +201,14 @@ export class GfCreateOrUpdateAccessDialogComponent implements OnInit {
   private async createAccess() {
     const filters = this.buildFilters();
 
+    const expiresAt: Date | null = this.accessForm.get('expiresAt')?.value;
+    const type: AccessType = this.accessForm.get('type')?.value;
+
     const access: CreateAccessDto = {
+      type,
       alias: this.accessForm.get('alias')?.value,
+      expiresAt:
+        type === 'API' && expiresAt ? expiresAt.toISOString() : undefined,
       filters: filters.length > 0 ? filters : undefined,
       granteeUserId: this.accessForm.get('granteeUserId')?.value,
       permissions: [this.accessForm.get('permissions')?.value]
@@ -209,8 +235,8 @@ export class GfCreateOrUpdateAccessDialogComponent implements OnInit {
           }),
           takeUntilDestroyed(this.destroyRef)
         )
-        .subscribe(() => {
-          this.dialogRef.close(access);
+        .subscribe((response) => {
+          this.dialogRef.close(response);
         });
     } catch (error) {
       console.error(error);
@@ -225,6 +251,8 @@ export class GfCreateOrUpdateAccessDialogComponent implements OnInit {
         this.holdings = getHoldingsForFilter(holdings);
 
         this.updateFiltersFormControl(this.data.access?.settings?.filters);
+
+        this.hasLoadedFilters = true;
 
         this.changeDetectorRef.markForCheck();
       });
@@ -241,7 +269,7 @@ export class GfCreateOrUpdateAccessDialogComponent implements OnInit {
 
     const access: UpdateAccessDto = {
       alias: this.accessForm.get('alias')?.value,
-      filters: filters.length > 0 ? filters : undefined,
+      filters: this.hasLoadedFilters ? filters : undefined,
       granteeUserId: this.accessForm.get('granteeUserId')?.value,
       id: accessId,
       permissions: [this.accessForm.get('permissions')?.value]