+++ libs/common/src/lib/interfaces/responses/create-access-response.interface.ts
@@ -1,0 +1,5 @@
+import { Access } from '@prisma/client';
+
+export interface CreateAccessResponse extends Omit<Access, 'hashedApiToken'> {
+  apiToken?: string;
+}